+++ TwoElectronics/source/PrintStruct.cpp
@@ -128,7 +128,7 @@ void PrintArray(double* array, size_t n, const char* FileName)
 }
 
 
-void Print_Count_Array(double* ee0_array,size_t * z_arr,size_t * zz_arr,int size,const char* file_name)
+void Print_Count_Array(double* ee0_array, unsigned long long* z_arr, unsigned long long* zz_arr, int size, const char* file_name)
 {
 	FILE* file;
 	file = fopen(file_name, "w");
